client/requirements.txt
pyserial
esptool